// front/src/app/models/IAutorisations.ts
export interface IAutorisations {
    ajouter_evenement: boolean;
    ajouter_perturbation: boolean;
    modifier_etat_perturbation_creation: boolean;
    gerer_contacts: boolean;
    gerer_organismes: boolean;
    gerer_contacts_preavis: boolean;
    gerer_contacts_avis_urgences: boolean;
    gerer_contacts_srb: boolean;
    gerer_utilisateurs: boolean;
    peut_deleguer: boolean;
}